+++ risker_read-it/public/app.html
@@ -4,20 +4,20 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>No more Risk! - Risk Management Dashboard</title>
-    <link rel="stylesheet" href="styles.css">
-    <script src="./libs/xlsx.full.min.js"></script>
-    <script src="logger.js"></script>
+    <link rel="stylesheet" href="../src/styles/styles.css">
+    <script src="../libs/xlsx.full.min.js"></script>
+    <script src="../src/scripts/logger.js"></script>
 
     <!-- Lade utils.js ZUERST, dann app.js -->
-    <script src="utils.js"></script>
-    <script src="./libs/chart.umd.js"></script>
-    <script src="app.js"></script>
+    <script src="../src/scripts/utils.js"></script>
+    <script src="../libs/chart.umd.js"></script>
+    <script src="../src/scripts/app.js"></script>
 </head>
 <body>
     <!-- Background Video -->
     <div class="background-video-container">
         <video id="background-video" autoplay muted loop playsinline>
-            <source src="background.mp4" type="video/mp4">
+            <source src="assets/background.mp4" type="video/mp4">
             Your browser does not support the video tag.
         </video>
         <div class="background-blur-overlay"></div>
