+++ risker_read-it/public/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <title>NO MORE RISK! Start</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="../src/styles/styles.css" />
     <style>
         html, body {
             height: 100%;
