+++ risker_read-it/public/riskmap.html
@@ -4,16 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>No more Risk! - RiskMap</title>
-    <link rel="stylesheet" href="styles.css">
-    <script src="./libs/xlsx.full.min.js"></script>
-    <script src="logger.js"></script>
-    <script src="utils.js"></script>
+    <link rel="stylesheet" href="../src/styles/styles.css">
+    <script src="../libs/xlsx.full.min.js"></script>
+    <script src="../src/scripts/logger.js"></script>
+    <script src="../src/scripts/utils.js"></script>
 </head>
 <body>
     <!-- Background Video -->
     <div class="background-video-container">
         <video id="background-video" autoplay muted loop playsinline>
-            <source src="background.mp4" type="video/mp4">
+            <source src="assets/background.mp4" type="video/mp4">
             Your browser does not support the video tag.
         </video>
         <div class="background-blur-overlay"></div>
