+++ risker_read-it/setup.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+npm install
+npx eslint -c ../.eslintrc.json .
+npx jest
