+++ risker_read-it/src/scripts/app.js
