+++ risker_read-it/src/scripts/utils.js
