--- setup.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-npm install
-npx eslint -c .eslintrc.json .
-npx jest
